requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0

# Data processing
pandas==2.2.0
requests==2.31.0

# AI Integration
anthropic==0.18.1

# CLI
click==8.1.7
rich==13.7.0

# Export formats
jinja2==3.1.3

# Utilities
python-dateutil==2.8.2
PyYAML==6.0.1
